requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
httpx==0.27.2
beautifulsoup4==4.12.3
anthropic==0.39.0
openai==1.51.0
python-dotenv==1.0.1
